+++ packages/ice/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v3.1.0
 
 - [fix] build data loader with correct browser list
+- [fix] process exit when build with errors on dev
+- [feat] optimize log info
 
 ## v3.0.6