+++ packages/ice/CHANGELOG.md
@@ -2,16 +2,18 @@
 
 ## v3.1.0
 
+- [feat] export suspense api
+- [feat] refactor ability of keep target code
+- [fix] support cli options of target, platform not is deprecated
 - [fix] build data loader with correct browser list
 - [fix] process exit when build with errors on dev
-- [feat] optimize log info
+- [fix] optimize log info
+- [fix] should not replace global env for ssr
 
 ## v3.0.6
 
-- [feat] export suspense api
 - [feat] support `import.meta.target`, `import.meta.renderer` and `import.meta.env`
 - [feat] support config `routes.injectInitialEntry` for memory router
-- [fix] should not replace global env for ssr
 - [fix] cli option `--platform` change to `--target`
 - [fix] remove code when dead code is `ObjectProperties`
 - [fix] rebuild server entry when document changed